Docker/web-service/app/requirements.txt
Flask==2.3.2
Werkzeug==2.3.4
psycopg2-binary==2.9.9
